+++ src/utils.ts
@@ -31,9 +31,9 @@ import {
   MarginAccount,
   ReferrerAlias,
   TradeEvent,
+  TradeEventV2,
   OpenOrdersMap,
 } from "./program-types";
-import { Network } from "./network";
 import * as types from "./types";
 import * as instructions from "./program-instructions";
 import { Decimal } from "./decimal";
@@ -502,7 +502,7 @@ export function convertDecimalToNativeInteger(amount: number): number {
  * does not divide perfectly by tick size (0.0001) if your order traded
  * against orders at different prices.
  */
-export function getTradeEventPrice(event: TradeEvent): number {
+export function getTradeEventPrice(event: TradeEvent | TradeEventV2): number {
   let decimalCostOfTrades = convertNativeBNToDecimal(event.costOfTrades);
   let decimalSize = convertNativeLotSizeToDecimal(event.size.toNumber());
   return decimalCostOfTrades / decimalSize;